move a few utilities to more localized packages
