+++ app/api/route/request.py
@@ -1,0 +1,6 @@
+from pydantic import BaseModel
+
+
+class BaseRequestModel(BaseModel):
+    class Config:
+        orm_mode = True
